requirements.txt
numpy>=1.26
pandas>=2.2
pyyaml>=6.0
pydantic>=2.7
jsonschema>=4.23
ruff>=0.6
pytest>=8.3
